refactored session components
