+++ src/index.css
@@ -25,3 +25,4 @@ code {
     'Courier New',
     monospace;
 }
+
