--- src/pages/SignUpPage.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const SignUpPage = () => (
-  <div>
-    <h1>SignUpPage</h1>
-  </div>
-);
-
-export default SignUpPage;
